refactor!: rename inputfield to input and textfield to inputfield (#1485)
* refactor(input)!: rename old inputfield component to input

* refactor(inputfield)!: rename old textfield to inputfield

* test(inputfield): match test suite description

* test(input): remove low value tests

* docs(input): clean input prop and sp
