+++ src/components/Input/Input.module.css
@@ -1,0 +1,13 @@
+@import '../../design-tokens/mixins.css';
+
+/*------------------------------------*\
+    # INPUT
+\*------------------------------------*/
+
+/**
+ * Default input styles 
+ */
+:where(.input) {
+  @mixin inputStyles;
+  padding-left: var(--eds-size-2);
+}
