+++ src/components/InputField/index.ts
@@ -1,2 +1,1 @@
 export { InputField as default } from './InputField';
-export type { InputFieldProps } from './InputField';
