--- src/components/TextField/TextField.test.ts
@@ -1,6 +1,0 @@
-import { generateSnapshots } from '@chanzuckerberg/story-utils';
-import * as TextFieldStoryFile from './TextField.stories';
-
-describe('<TextField />', () => {
-  generateSnapshots(TextFieldStoryFile);
-});
