--- src/components/TextField/index.ts
@@ -1,1 +1,0 @@
-export { TextField as default } from './TextField';
